% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipType.R
\name{chipType}
\alias{chipType}
\title{Determine ChIP-Seq Type of Protein Complex Binding}
\usage{
chipType(count, cutoff = 50L, fold = 10, h = 0.1, plot = TRUE)
}
\arguments{
\item{count}{A two-column matrix of read counts or a SummarizedExperiment,
where columns are samples and rows are peaks or high coverage bins. This
object can be generated by function \code{regionReads}.}

\item{cutoff}{A numeric cut off on \code{count} matrix. If positive, only
peaks/bins with counts larger than \code{cutoff} in at least one sample are
used to estimate the binding type. We recommend a larger cutoff since
background signal can dramatically mask the right estimation of kernal
density, especially for deep sequenced ChIP-seq samples. (Default: 50)}

\item{fold}{A numeric threshold to help determine the binding type. In
detail, if top two estimated modes on smoothed kernal density have a height
differece less than the folds given by \code{fold}, binding type will be
determined as bimodel; otherwise, it is unimodel. This number should be
larger than 1. (Default: 10)}

\item{h}{Initial smoothing factor when estimating kernal density of raw fold
changes for bump hunting. (Default: 0.1)}

\item{plot}{A logical indicator that if M-A plot and smoothed kernal density
should be visualized. (Default: TRUE)}
}
\value{
A character with value either "bimodel" or "unimodel" to represent estimated
binding type.
}
\description{
Estimate the types of protein complex binding. Protein complex
binding might act similarly to normal transcription factors, where the
changes are symmetrical between two biological conditions (unimodel on fold
changes); or the changes might be globally one-side accompanied with some
non-changing bindings (bimodel on fold changes). This function help to
determine the binding type of given ChIP-Seq samples from two biological
conditions using kernal density information of raw fold changes. As this
function was designed to work on the raw counts (no normalization needed),
only one replicate from each condition is allowed (input a two-column count
matrix); otherwise, coverage difference acorss replicates might bias
determination.
}
\examples{
## load sample data
data(complex)
names(complex)

## test sample data
chipType(count=complex$counts)
}

