% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\docType{data}
\name{complex}
\alias{complex}
\title{Example of Protein Complex Data}
\format{A list with two conponents:
\describe{
  \item{counts}{read counts}
  \item{bins}{300 bp bins}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE67382}
}
\description{
This data contains two compared samples that measures the DNA binding
differences of protein complex. ChIP-seq reads are counted for 300bp bins.
Only data from chromosome 10 is included.
}

