% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexDiff.R
\name{complexDiff}
\alias{complexDiff}
\title{Differential Binding Estimation for Protein Complexes}
\usage{
complexDiff(count, bins, meta, design, sizefac, rccut = 15, fccut = 0.4,
  gap = 2, diffmeth = c("DESeq2", "limma", "ttest"))
}
\arguments{
\item{count}{A matrix of read counts for genome-wide bins,
where columns correspond to samples and rows correspond to bins.
This matrix can be generated by function \code{regionCounts}.}

\item{bins}{A GRanges object recording bins corresponding to
rows of \code{count} matrix.
This object can be generated by function \code{regionCounts}.}

\item{meta}{A DataFrame object recording sample annotations.
Rows of \code{meta} correspond to columns of \code{count}.}

\item{design}{A formula object which expresses how the counts for
each bin depend on the variables in \code{meta}, e.g. '~ group +
condition' etc. By default, the last variable in the formula
will be used to build the differential binding contrast. At most
two variables are allowed if \code{diffmeth} is set to 'ttest'.
(Details see below)}

\item{sizefac}{A numeric vector indicating estimated size of samples
for normalization purpose. This vector can be generated by
function \code{complexNorm}.}

\item{rccut}{A numeric cutoff on normalized \code{count} matrix by
\code{sizefac}. If positive, only bins with normalized counts larger than
\code{rccut} in at least one sample are selected for fold change estimate.
Unlike other functions in this package, moderate cutoff would be better
as too large results more false negative and too small increases the
time cost. (Default: 15)}

\item{fccut}{A numeric cutoff on bins' smoothed log2foldchanges for bump
hunting. Neighbor bins with fold change larger than this value will be
merged together with allowed gaps. (Default: 0.4)}

\item{gap}{A integer specifying the gaps allowed for bin merging, in the
unit of number of bins. (Default: 2)}

\item{diffmeth}{Method for statistical testing of differential binding.
(Default: 'DESeq2')}
}
\value{
A list with the following components:
\item{region}{A GRanges object containing potential regions with
differential binding}
\item{diff}{A DataFrame containing estimated significance of
potential differential binding regions, including tested statistic,
log2 fold change and p-value}
}
\description{
This is the main function of this package. It accepts
outputs of other functions in this package, and integrates available
strategies of signal smoothing, bump hunting and differential testing
from other packages, and reports differential binding regions with
estimated significance. It is important that the inputs are genome-
wide bin level read counts, instead of peak level. Also, it is noted
that each read should be only assigned to one bin if multiple
overlapping exists, as done by function \code{regionCounts}.
}
\details{
If 'ttest' is selected, ...
}
\examples{
## load sample data
data(complex)
names(complex)

## test sample data
sizefac <- complexNorm(complex$counts,plot=TRUE)
meta <- data.frame(cond=c("ctr","tre"))
cdiff <- complexDiff(complex$counts,complex$bins,meta,design=~cond,sizefac)

## return values
names(cdiff)
cdiff$region
hist(width(cdiff$region),nclass=30,xlab="region width",
main="Width of potential differential regions")
hist(-log10(cdiff$diff$pvalue),nclass=30,xlab="-log10 pvalue",
main="Estimated significance")
}

