% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexDiffPerm.R
\name{complexDiffPerm}
\alias{complexDiffPerm}
\title{Differential Binding Estimation with Permutation}
\usage{
complexDiffPerm(count, bins, meta, design, sizefac, rccut = 15, fccut = 0.4,
  gap = 2, permute = NULL, maxperm = 10, diffmeth = c("DESeq2", "limma",
  "ttest"))
}
\arguments{
\item{count}{A matrix of read counts for genome-wide bins,
where columns correspond to samples and rows correspond to bins.
This matrix can be generated by function \code{regionCounts}.}

\item{bins}{A GRanges object recording bins corresponding to
rows of \code{count} matrix.
This object can be generated by function \code{regionCounts}.}

\item{meta}{A DataFrame object recording sample annotations.
Rows of \code{meta} correspond to columns of \code{count}.}

\item{design}{A formula object which expresses how the counts for
each bin depend on the variables in \code{meta}, e.g. '~ group +
condition' etc. By default, the last variable in the formula
will be used to build the differential binding contrast. At most
two variables are allowed if \code{diffmeth} is set to 'ttest'.
(Details see below)}

\item{sizefac}{A numeric vector indicating estimated size of samples
for normalization purpose. This vector can be generated by
function \code{complexNorm}.}

\item{rccut}{A numeric cutoff on normalized \code{count} matrix by
\code{sizefac}. If positive, only bins with normalized counts larger than
\code{rccut} in at least one sample are selected for fold change estimate.
Unlike other functions in this package, moderate cutoff would be better
as too large results more false negative and too small increases the
time cost. (Default: 15)}

\item{fccut}{A numeric cutoff on bins' smoothed log2foldchanges for bump
hunting. Neighbor bins with fold change larger than this value will be
merged together with allowed gaps. (Default: 0.4)}

\item{gap}{A integer specifying the gaps allowed for bin merging, in the
unit of number of bins. (Default: 2)}

\item{permute}{A matrix where each row contains shuffled indices of
samples. If \code{NULL}, permutation is automatically obtained by shuffling
samples between compared conditions. This object can be generated by using
R package \code{permute} as well.}

\item{maxperm}{Maximum number of permutations to be finished. (Default:10)}

\item{diffmeth}{Method for statistical testing of differential binding.
(Default: 'DESeq2')}
}
\value{
A list with the following components:
\item{region}{A GRanges object containing potential regions with
differential binding}
\item{diff}{A DataFrame containing estimated significance of
potential differential binding regions, including tested statistic,
log2 fold change, p-value and permutation p-value}
}
\description{
This function does differential binding estimation and additional
permutation analysis. This function works only if multiple replicates
are provided for comparison. For differential binding estimation, see
function \code{complexDiff}. Permutation analysis is obtained by
shuffling samples between compared conditions. Customized shuffling
can also be specified by parameter \code{permute} for complicated
experiment designs.
}
\details{
Three methods are provided for significance estimation of differential
binding. \code{DESeq2} allows pseudo-estimation for comparisons without
replicates; otherwise, all methods can be used for comparisons with at
least two replicates. The \code{design} formula can be specified as
suggested by \code{DESeq2} and \code{limma} if these two methods are
selected. For \code{ttest}, \code{design} can either contain one or two
components, referring to student's t-test or paired t-test based on
logarithm scaled data. For consistance with other packages, the last
component in \code{design} formula is the contrast on which the final
differential estimation are reported.

The default permutation works on the main contrast only (the last
component in \code{design} formula). If customized permutation needed,
e.g. invloving several components in \code{design} formula, users should
provided the indecies of permutations through parameter \code{permute}.
}

