% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexNorm.R
\name{complexNorm}
\alias{complexNorm}
\title{Scaling Factor Estimates Based on Kernal Density Bump Hunting}
\usage{
complexNorm(count, cutoff = 50L, fold = 10, h = 0.1, plot = FALSE,
  sanity = FALSE, samfac = NULL)
}
\arguments{
\item{count}{A matrix of read counts for peaks or
high coverage bins, where columns are samples and rows are peaks
or bins. This matrix can be generated by function \code{regionCounts}.}

\item{cutoff}{See function \code{complexNormPair}. (Default: 50)}

\item{fold}{See function \code{complexNormPair}. (Default: 10)}

\item{h}{See function \code{complexNormPair}. (Default: 0.1)}

\item{plot}{See function \code{complexNormPair}. (Default: FALSE)}

\item{sanity}{A logical indicator if checking sanity across replicates
in the same conditions. A negative report of sanity check indicates
either a bad experiment or a bad initiation of function parameters.
(Default: FALSE)}

\item{samfac}{\code{NULL} or a two-level factor specifying biological
conditions for samples in \code{count} (e.g. control
& treatment). This parameter is only appliable when \code{sanity} is
TRUE. (Default: NULL)}
}
\value{
A numeric vector indicating estimated size of samples.
}
\description{
This function iteratively invokes \code{complexNormPair}
for scaling factor estimates across multiple samples, i.e. replicates
and conditions.
}
\examples{
## load sample data
data(complex)
names(complex)

## test sample data
complexNorm(complex$counts)
complexNorm(complex$counts,sanity=TRUE,samfac=factor(c("ctr","tre")))
}

