% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexType.R
\name{complexType}
\alias{complexType}
\title{Type Determination of Protein Complex Binding}
\usage{
complexType(count, cutoff = 50L, fold = 10, h = 0.1, plot = TRUE)
}
\arguments{
\item{count}{A two-column matrix of read counts for peaks or
high coverage bins, where columns are samples and rows are peaks
or bins. This matrix can be generated by function \code{regionCounts}.}

\item{cutoff}{A numeric cut off on \code{count} matrix. If positive,
only peaks/bins with counts larger than \code{cutoff} in at least one
sample are used to estimate the binding type. We recommend a larger
cutoff since background signal can dramatically mask the right estimation
of kernal density, especially for deep sequenced ChIP-seq samples.
(Default: 50)}

\item{fold}{A numeric threshold to help determining the binding type.
In detail, if top 2 estimated modes on smoothed kernal density have
a height differece less than the folds given by \code{fold}, binding
type will be determined as bimodel; otherwise, it is unimodel. This
number should be larger than 1. (Default: 10)}

\item{h}{Initial smoothing factor when estimating kernal density for bump
hunting. (Default: 0.1)}

\item{plot}{A logical indicator that if MA plot and smoothed kernal density
should be visualized. (Default: TRUE)}
}
\value{
A character with value either "bimodel" or "unimodel",
indicating the binding types.
}
\description{
Estimate the types of protein complex binding. Protein
complex binding might act similarly to normal transcription factors,
where normally changes are symmetrical between two biological
conditions (unimodel on fold change); or the changes might be
globally one-side accompanied with some non-changing bindings (bimodel
on fold change). This function can help to estimate the binding type
given ChIP-Seq samples from two biological conditions. As this function
was designed to work on the very raw count data (no normalization needed),
only one replicate from each condition is allowed (total two columns in
the count matrix); otherwise, coverage difference acorss replicates
might bias determination.
}
\examples{
## load sample data
data(complex)
names(complex)

## test sample data
complexType(complex$counts)
}

