% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffRegions.R
\name{diffRegions}
\alias{diffRegions}
\title{Differential Binding Estimation for Protein Complexes}
\usage{
diffRegions(count, bins, meta, design, sizefac, rccut = 15, fccut = 0.4,
  gap = 2, diffmeth = c("DESeq2", "limma", "ttest"))
}
\arguments{
\item{count}{A matrix of read counts for genome-wide bins,
where columns correspond to samples and rows correspond to bins.
This matrix can be generated by function \code{regionReads}.}

\item{bins}{A GRanges object recording bins corresponding to
rows of \code{count} matrix.
This object can be generated by function \code{regionReads}.}

\item{meta}{A DataFrame object recording sample annotations.
Rows of \code{meta} correspond to columns of \code{count}.}

\item{design}{A formula object which expresses how the counts for
each bin depend on the variables in \code{meta}, e.g. '~ group +
condition' etc. By default, the last variable in the formula
will be used to build the differential binding contrast. At most
two variables are allowed if \code{diffmeth} is set to 'ttest'.
(Details see below)}

\item{sizefac}{A numeric vector indicating estimated size of samples
for normalization purpose. This vector can be generated by
function \code{sizeFac}.}

\item{rccut}{A numeric cutoff on normalized \code{count} matrix using
\code{sizefac}. If positive, only bins with normalized counts larger than
\code{rccut} in at least one sample are selected for fold change estimate.
Unlike other functions in this package, moderate cutoff would be better
as too large results more false negative and too small increases the
time cost. (Default: 15)}

\item{fccut}{A numeric cutoff on smoothed log2foldchanges of bins for
bump hunting of differtial binding regions. Neighbor bins with fold change
larger than this value will be merged together with allowed gaps.
(Default: 0.4)}

\item{gap}{A integer specifying the gaps allowed for bin merging, in the
unit of number of bins. (Default: 2)}

\item{diffmeth}{Method for statistical testing of differential binding.
(Default: 'DESeq2')}
}
\value{
A GRanges object containing potential regions with
differential binding, as well as statistical significances as meta columns.
}
\description{
This is the main function of this package. It accepts
outputs of other functions in this package, and integrates available
strategies of signal smoothing, bump hunting and differential testing
from other packages, and reports differential binding regions with
estimated significance. It is important that the inputs are genome-
wide bin level read counts, instead of peak level. Also, it is noted
that each read should be only assigned to one bin if multiple
overlapping exists, as done by function \code{regionReads}.
}
\details{
Three methods are provided for significance estimation of differential
binding. \code{DESeq2} allows pseudo-estimation for comparisons without
replicates; otherwise, all methods can be used for comparisons with at
least two replicates. The \code{design} formula can be specified as
suggested by \code{DESeq2} and \code{limma} if these two methods are
selected. For \code{ttest}, \code{design} can either contain one or two
components, referring to student's t-test or paired t-test based on
logarithm scaled data. For consistance with other packages, the last
component in \code{design} formula is the contrast on which the final
differential estimation are reported.
}
\examples{
## load sample data
data(complex)
names(complex)

## test sample data
sizefac <- sizeFac(complex$counts,plot=TRUE)$sizefac
meta <- data.frame(cond=c("ctr","tre"))
dr <- diffRegions(complex$counts,complex$bins,meta,design=~cond,sizefac)

## return values
dr
hist(width(dr),nclass=30,xlab="region width",
     main="Width of potential differential regions")
hist(-log10(dr$pvalue),nclass=30,xlab="-log10 pvalue",
     main="Estimated significance")
}

