% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFac.R
\name{sizeFac}
\alias{sizeFac}
\title{Estimate Size Factors of Samples}
\usage{
sizeFac(count, cutoff = 50L, fold = 10, h = 0.1, plot = FALSE,
  sanity = FALSE, samfac = NULL)
}
\arguments{
\item{count}{A matrix of read counts for peaks or
high coverage bins, where columns are samples and rows are peaks
or bins. This matrix can be generated by function \code{regionReads}.}

\item{cutoff}{A numeric cut off on \code{count} matrix. If positive,
only peaks/bins with counts larger than \code{cutoff} in at least one
sample are used to estimate the scaling factor. We recommend a larger
cutoff since background signal can dramatically mask the right estimation
of kernal density, especially for deep sequenced ChIP-seq samples.
(Default: 50)}

\item{fold}{A numeric threshold to help determining the binding type.
In detail, if top 2 estimated modes on smoothed kernal density have
a height differece less than the folds given by \code{fold}, binding
type will be determined as bimodel; otherwise, it is unimodel. This
number should be larger than 1. (Default: 10)}

\item{h}{Initial smoothing factor when estimating kernal density for
bump hunting. (Default: 0.1)}

\item{plot}{A logical indicator that if M-A plot and smoothed kernal
density should be visualized. (Default: FALSE)}

\item{sanity}{A logical indicator if checking sanity across replicates
in the same conditions. A negative report of sanity check indicates
either a bad experiment (e.g. binding type is not consistent across
replicates) or a bad initiation of function parameters (e.g.
\code{cutoff} and \code{fold} are not pre-estimated well). However,
a negative report of sanity check doesn't neccessarily mean a bad
estimation of size factors, as the strategy of hunting kernal density
mode can mostly find the right scaling regardless of unimodel or
bimodel. (Default: FALSE)}

\item{samfac}{\code{NULL} or a two-level factor specifying biological
conditions for samples in \code{count} (e.g. control
& treatment). This parameter is only appliable when \code{sanity} is
TRUE. (Default: NULL)}
}
\value{
A list with the following conponents:
\item{sizefac}{A numeric vector indicating estimated size factors
of samples}
\item{type}{A character vector with value either "bimodel" or "unimodel",
indicating the binding types by comparing ro the sample "control"}
}
\description{
This function estimates sample size factors for normalization
purpose in downsteam analysis. Size factors of sample pairs are
estimated firstly by comparing samples to the same sample (i.e. the
sample corresponding to first column of count matrix). Then, size factors
are combined across all samples with the median size factor as 1.
In detail, binding type is first estimated using the same strategy as
function \code{chipType} for each sample pairs. When biomodel, size factor
is calcualted based on a decision on which kernale density mode to be used
for scaling.
}
\examples{
## load sample data
data(complex)
names(complex)

## test sample data
sizeFac(complex$counts)
}

