% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionReads.R
\name{regionReads}
\alias{regionReads}
\title{Count ChIP-Seq Reads for Genomic Regions}
\usage{
regionReads(samples, peaks = NULL, binsize = 300L, bincut = 0L,
  readlen = NULL, read2pos = "5", ignoreDup = TRUE)
}
\arguments{
\item{samples}{A vector of paths to ChIP-seq files in BAM or Bigwig
format.}

\item{peaks}{GRanges or \code{NULL}. If \code{NULL}, genome-wide
high coverage bins with length \code{binsize} will be pre-selected as
potential peaks; otherwise, ChIP-seq peaks should be provided as a
GRanges object, e.g. a union of peaks identified by peak caller
from all samples. (Default: NULL)}

\item{binsize}{A number specify the bin size. This parameter is
only appliable when \code{peaks=NULL}. (Default: 300)}

\item{bincut}{A numeric cut off on read counts to nominate bins as
high coverage candidates. This parameter is only appliable when
\code{peaks=NULL}. (Default: 0)}

\item{readlen}{Sample fragment length with which BIGWIG files created.
The signal in BIGWIG file is \code{readlen} times of original reads
number. One way to count original reads is to privde \code{readlen} here.
This parameter is only appliable when \code{samples} are BIGWIG files.
If \code{NULL}, it is recommended to multiply \code{bincut} by a scale
factor (normally fragment length). (Default: NULL)}

\item{read2pos}{Parameter of function \code{featureCounts} in package
Rsubread. This parameter is only appliable when
\code{samples} are bam files. (Default: '5')}

\item{ignoreDup}{Parameter of function \code{featureCounts} in package
Rsubread. This parameter is only appliable when
\code{samples} are bam files. (Default: TRUE)}
}
\value{
A list with the following components:
\item{count}{A data matrix containing read counts for each region in
each sample}
\item{regions}{A GRanges object recording regions of each row in the
count matrix}
}
\description{
Count reads for customized regions or genome wide
bins for multiple ChIP-Seq samples in BAM (.bam) or BIGWIG
(.bw or .bigwig) format. When no customized peaks provided, this
function will generate bin-level counts based on genome
information embedded in BAM or BIGWIG files. For BAM files,
counts will be generated using Rsubread package; for BIGWIG
files, counts will be generated by function in rtracklayer package.
}
\examples{
library(GenomicRanges)
bams <- c(system.file("extdata", "control.bam", package="ComplexDiff"),
          system.file("extdata", "treated.bam", package="ComplexDiff"))
bins <- GRanges("chr1",IRanges(start = seq(1000000,2000000,300),
                      end = seq(1000000,2000000,300)+300-1))
rc <- regionReads(bams,peaks=bins)
names(rc)

}

